refactored commons package structure
