+++ ChathamRoboticsCommon/src/main/java/org/chathamrobotics/common/hardware/utils/HardwareListener.java
@@ -1,4 +1,4 @@
-package org.chathamrobotics.common.utils.hardware;
+package org.chathamrobotics.common.hardware.utils;
 
 import android.util.Log;
