+++ Team9853/src/main/java/org/firstinspires/ftc/team9853/Robot9853.java
@@ -6,7 +6,7 @@ import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.Gamepad;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
-import org.chathamrobotics.common.Robot;
+import org.chathamrobotics.common.robot.Robot;
 import org.chathamrobotics.common.systems.HolonomicDriver;
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.team9853.systems.GlyphGripper;
