+++ Team9853/src/main/java/org/firstinspires/ftc/team9853/opmodes/teleop/CompDrive.java
@@ -3,7 +3,7 @@ package org.firstinspires.ftc.team9853.opmodes.teleop;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import org.chathamrobotics.common.utils.hardware.Controller;
+import org.chathamrobotics.common.Controller;
 import org.firstinspires.ftc.team9853.Robot9853;
 
 /**
