+++ Team9853/src/main/java/org/firstinspires/ftc/team9853/opmodes/teleop/JewelTester.java
@@ -2,8 +2,8 @@ package org.firstinspires.ftc.team9853.opmodes.teleop;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 
-import org.chathamrobotics.common.utils.hardware.Controller;
-import org.chathamrobotics.common.utils.hardware.IsBusyException;
+import org.chathamrobotics.common.Controller;
+import org.chathamrobotics.common.IsBusyException;
 import org.firstinspires.ftc.team9853.Robot9853;
 import org.firstinspires.ftc.team9853.systems.JewelDisplacer;
