+++ Team9853/src/main/java/org/firstinspires/ftc/team9853/opmodes/teleop/ServoTester.java
@@ -6,7 +6,7 @@ import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.Gamepad;
 import com.qualcomm.robotcore.hardware.Servo;
 
-import org.chathamrobotics.common.utils.robot.RobotLogger;
+import org.chathamrobotics.common.robot.RobotLogger;
 
 import java.util.Map;
