+++ Team9853/src/main/java/org/firstinspires/ftc/team9853/systems/JewelDisplacer.java
@@ -4,11 +4,11 @@ import com.qualcomm.robotcore.hardware.ColorSensor;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.hardware.Servo;
 
-import org.chathamrobotics.common.Robot;
+import org.chathamrobotics.common.robot.Robot;
 import org.chathamrobotics.common.systems.RackAndPinion;
-import org.chathamrobotics.common.utils.hardware.HardwareListener;
-import org.chathamrobotics.common.utils.hardware.IsBusyException;
-import org.chathamrobotics.common.utils.robot.RobotLogger;
+import org.chathamrobotics.common.hardware.utils.HardwareListener;
+import org.chathamrobotics.common.IsBusyException;
+import org.chathamrobotics.common.robot.RobotLogger;
 
 import java.util.Timer;
